+++ lib/hexapdf/type/annotations/widget.rb
@@ -48,6 +48,8 @@ module HexaPDF
       # See: PDF2.0 s12.5.6.19, HexaPDF::Type::Annotation
       class Widget < Annotation
 
+        include BorderStyling
+
         # The dictionary used by the /MK key of the widget annotation.
         class AppearanceCharacteristics < Dictionary
 
@@ -119,102 +121,6 @@ module HexaPDF
             color = Content::ColorSpace.device_color_from_specification(color)
             (self[:MK] ||= {})[:BG] = color.components
             self
-          end
-        end
-
-        # Describes the border of an annotation.
-        #
-        # The +color+ property is either +nil+ if the border is transparent or else a device color
-        # object - see HexaPDF::Content::ColorSpace.
-        #
-        # The +style+ property can be one of the following:
-        #
-        # :solid::      Solid line.
-        # :beveled::    Embossed rectangle seemingly raised above the surface of the page.
-        # :inset::      Engraved rectangle receeding into the page.
-        # :underlined:: Underlined, i.e. only the bottom border is draw.
-        # Array:        Dash array describing how to dash the line.
-        BorderStyle = Struct.new(:width, :color, :style, :horizontal_corner_radius,
-                                 :vertical_corner_radius)
-
-        # :call-seq:
-        #   widget.border_style                                      => border_style
-        #   widget.border_style(color: 0, width: 1, style: :solid)   => widget
-        #
-        # Returns a BorderStyle instance representing the border style of the widget when no
-        # argument is given. Otherwise sets the border style of the widget and returns self.
-        #
-        # When setting a border style, arguments that are not provided will use the default: a
-        # border with a solid, black, 1pt wide line. This also means that multiple invocations will
-        # reset *all* prior values.
-        #
-        # +color+:: The color of the border. See
-        #           HexaPDF::Content::ColorSpace.device_color_from_specification for information on
-        #           the allowed arguments.
-        #
-        #           If the special value +:transparent+ is used when setting the color, a
-        #           transparent is used. A transparent border will return a +nil+ value when getting
-        #           the border color.
-        #
-        # +width+:: The width of the border. If set to 0, no border is shown.
-        #
-        # +style+:: Defines how the border is drawn. can be one of the following:
-        #
-        #           +:solid+::      Draws a solid border.
-        #           +:beveled+::    Draws a beveled border.
-        #           +:inset+::      Draws an inset border.
-        #           +:underlined+:: Draws only the bottom border.
-        #           Array::         An array specifying a line dash pattern (see
-        #                           HexaPDF::Content::LineDashPattern)
-        def border_style(color: nil, width: nil, style: nil)
-          if color || width || style
-            color = if color == :transparent
-                      []
-                    else
-                      Content::ColorSpace.device_color_from_specification(color || 0).components
-                    end
-            width ||= 1
-            style ||= :solid
-
-            (self[:MK] ||= {})[:BC] = color
-            bs = self[:BS] = {W: width}
-            case style
-            when :solid then bs[:S] = :S
-            when :beveled then bs[:S] = :B
-            when :inset then bs[:S] = :I
-            when :underlined then bs[:S] = :U
-            when Array
-              bs[:S] = :D
-              bs[:D] = style
-            else
-              raise ArgumentError, "Unknown value #{style} for style argument"
-            end
-            self
-          else
-            result = BorderStyle.new(1, nil, :solid, 0, 0)
-            if (ac = self[:MK]) && (bc = ac[:BC]) && !bc.empty?
-              result.color = Content::ColorSpace.prenormalized_device_color(bc.value)
-            end
-
-            if (bs = self[:BS])
-              result.width = bs[:W] if bs.key?(:W)
-              result.style = case bs[:S]
-                             when :S then :solid
-                             when :B then :beveled
-                             when :I then :inset
-                             when :U then :underlined
-                             when :D then bs[:D].value
-                             else :solid
-                             end
-            elsif key?(:Border)
-              border = self[:Border]
-              result.horizontal_corner_radius = border[0]
-              result.vertical_corner_radius = border[1]
-              result.width = border[2]
-              result.style = border[3] if border[3]
-            end
-
-            result
           end
         end